pyproject.toml
[tool.poetry]
name = "Codebox"
version = "0.1.0"
description = ""
authors = ["André Felipe Dias <andref.dias@gmail.com>"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
fastapi = "*"
hypercorn = "*"
loguru = "*"
orjson = "*"
pydantic = "*"
python-dotenv = "*"
pygments = "*"
python = "^3.11"
stackprinter = "*"
uvloop = "*"

[tool.poetry.group.dev.dependencies]
asgi-lifespan = "*"
alt-pytest-asyncio = "*"
flaky = "*"
httpx = { version = "*", extras = ["cli"] }
mypy = "*"
pip-audit = "*"
pytest = "*"
pytest-cov = "*"
ruff = "*"
types-orjson = "*"

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]
default_async_timeout = 600

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 100
select = [
    "ARG",
    "B",
    "C4",
    "E",
    "ERA",
    "F",
    "I",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "RET",
    "S",
    "T10",
    "UP",
    "W",
]
ignore = ["B008", "S603"]
target-version = "py311"

[tool.ruff.format]
quote-style = "single"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**" = ["ARG", "PLR2004", "S"]
"smoke_test.py" = ["S"]
